+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ProductModel } from '../../models/product.model';
 import { ProductsService } from '../../services/products.service';
+import { CartService } from '../../../cart/services/cart.service';
 
 @Component({
   selector: 'app-product-list',
@@ -10,9 +11,13 @@ import { ProductsService } from '../../services/products.service';
 export class ProductListComponent implements OnInit {
 
   books: ProductModel[];
-  constructor(private productService: ProductsService) { }
+  constructor(private productService: ProductsService, private cartService: CartService) { }
 
   ngOnInit() {
     this.books = this.productService.getBooks();
+  }
+
+  onAddedToCart(product: ProductModel): void {
+    this.cartService.addToCart(product);
   }
 }