+++ changelog.md
@@ -1,6 +1,11 @@
 ## 10.07
 
 Created initial project structure.
-Added Product, ProductList Cart components, Product and Cart services, Product model. All of the are used in AppComponent. 
+Added Product, ProductList Cart components, Product and Cart services, Product model. All of the are used in AppComponent.
+
+## 20.07
+
+Updated project tot use Angilar 8.
+Refactored Cart to show count and sum for added items. Added some minimal styling via component styling, added hover styling.